+++ src/morphocut/core.py
@@ -12,7 +12,7 @@ from typing import (
     Callable,
     Dict,
     Generic,
-    Iterable,
+    Iterator,
     List,
     Optional,
     Tuple,
@@ -31,7 +31,7 @@ class StreamTransformer(ABC):
         self.id = "{:x}".format(id(self))
 
     @abstractmethod
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: "Stream") -> "Stream":
         while False:
             yield
 
@@ -370,14 +370,14 @@ class Variable(Generic[T]):
 RawOrVariable = Union[T, Variable[T]]
 NodeCallReturnType = Union[None, Variable, Tuple[Variable]]
 
-Stream = Iterable["StreamObject"]
-r"""A stream is an Iterable of :py:class:`StreamObject`\ s."""
+Stream = Iterator["StreamObject"]
+r"""A stream is an Iterator of :py:class:`StreamObject`\ s."""
+
 
 
 class EmptyPipelineStackError(Exception):
     """Raised when a node is created outside of a Pipeline context."""
 
-    pass
 
 
 class Node(StreamTransformer):
@@ -439,8 +439,11 @@ class Node(StreamTransformer):
             _resolve_variable(obj, v) for v in (getattr(self, n) for n in names)
         )
 
-    def prepare_output(self, obj, *values):
+    def prepare_output(self, obj, *values, n_remaining_hint=None):
         """Update obj using the values corresponding to the output ports."""
+
+        if n_remaining_hint is not None:
+            obj.n_remaining_hint = n_remaining_hint
 
         if not self.outputs:
             if any(v is not None for v in values):
@@ -646,16 +649,20 @@ class StreamObject(abc.MutableMapping):
     """
     An object in the :py:obj:`Stream` that wraps all values.
 
+    Attributes:
+        data (dict): The data associated with this stream object.
+        n_remaining_hint (int, optional): Approximate number of remaining objects in the stream including the current object.
+
     A value can be retrieved by indexing: ``obj[var]``
     """
 
-    __slots__ = ["data", "stream_length"]
+    __slots__ = ["data", "n_remaining_hint"]
 
-    def __init__(self, data: Dict = None, stream_length=None):
+    def __init__(self, data: Dict = None, n_remaining_hint: Optional[int] = None):
         if data is None:
             data = {}
         self.data = data
-        self.stream_length = stream_length
+        self.n_remaining_hint = n_remaining_hint
 
     def copy(self) -> "StreamObject":
         """Create a shallow copy."""
@@ -691,6 +698,14 @@ class StreamObject(abc.MutableMapping):
             raise ValueError("No names were supplied")
 
         return {k: self[v] for k, v in kwargs.items()}
+
+
+def check_stream(stream: Optional[Stream]) -> Stream:
+    """Ensure that `stream` is a valid stream."""
+
+    if stream is None:
+        return iter([StreamObject(n_remaining_hint=1)])
+    return stream
 
 
 class Pipeline(StreamTransformer):
@@ -765,8 +780,8 @@ class Pipeline(StreamTransformer):
         Returns:
             Stream: An iterable of stream objects.
         """
-        
-        stream = self.check_stream(stream)
+
+        stream = check_stream(stream)
 
         # Here, the stream is not automatically closed,
         # as this would happen instantaneously.