+++ src/morphocut/stream.py
@@ -4,7 +4,8 @@ import itertools
 import pprint
 from queue import Queue
 from threading import Thread
-from typing import Callable, Iterable, Optional, Tuple, Union
+from typing import Callable, Collection, Iterable, Optional, Tuple, Union
+from morphocut.stream_estimator import StreamEstimator
 
 import tqdm
 from deprecated.sphinx import deprecated
@@ -64,15 +65,15 @@ class Progress(Node):
             if self.monitor_interval is not None:
                 progress.monitor_interval = self.monitor_interval
 
-            for obj in progress:
+            for n_processed, obj in enumerate(progress):
 
                 description = self.prepare_input(obj, "description")
 
                 if description:
                     progress.set_description(description)
 
-                if obj.stream_length is not None:
-                    progress.total = obj.stream_length
+                if obj.n_remaining_hint is not None:
+                    progress.total = n_processed + obj.n_remaining_hint
 
                 yield obj
 
@@ -80,6 +81,9 @@ class Progress(Node):
 @deprecated(reason="Deprecated in favor of Progress.", version="0.2.x")
 def TQDM(*args, **kwargs):
     return Progress(*args, **kwargs)
+
+
+TQDM.__doc__ = Progress.__doc__
 
 
 @ReturnOutputs
@@ -99,7 +103,7 @@ class Slice(Node):
         super().__init__()
         self.args = args
 
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: Stream):
         with closing_if_closable(stream):
             for obj in itertools.islice(stream, *self.args):
                 if obj.stream_length is not None:
@@ -109,7 +113,6 @@ class Slice(Node):
                 yield obj
 
 
-@ReturnOutputs
 class StreamBuffer(Node):
     """
     Buffer the stream.
@@ -126,7 +129,7 @@ class StreamBuffer(Node):
         super().__init__()
         self.queue = Queue(maxsize)
 
-    def _fill_queue(self, stream):
+    def _fill_queue(self, stream: Stream):
         try:
             with closing_if_closable(stream):
                 for obj in stream:
@@ -134,7 +137,7 @@ class StreamBuffer(Node):
         finally:
             self.queue.put(self._sentinel)
 
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: Stream):
         thread = Thread(target=self._fill_queue, args=(stream,), daemon=True)
         thread.start()
 
@@ -159,11 +162,11 @@ class PrintObjects(Node):
        *args (Variable): Variables to display.
     """
 
-    def __init__(self, *args: Tuple[Variable]):
+    def __init__(self, *args: Variable):
         super().__init__()
         self.args = args
 
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: Stream):
         with closing_if_closable(stream):
             for obj in stream:
                 print("Stream object at 0x{:x}".format(id(obj)))
@@ -190,7 +193,7 @@ class Enumerate(Node):
         super().__init__()
         self.start = start
 
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: Stream):
         with closing_if_closable(stream):
             for i, obj in enumerate(stream, start=self.start):
                 yield self.prepare_output(obj, i)
@@ -200,12 +203,12 @@ class Enumerate(Node):
 @Output("value")
 class Unpack(Node):
     """
-    |stream| Unpack values from an iterable into the :py:obj:`~morphocut.core.Stream`.
+    |stream| Unpack values from a collection into the :py:obj:`~morphocut.core.Stream`.
 
     The result is basically the cross-product of the stream with the iterable.
 
     Args:
-        iterable (Iterable or Variable): An iterable to unpack.
+        collection (Collection or Variable): An iterable to unpack.
 
     Returns:
        Variable: One value from the iterable.
@@ -227,19 +230,24 @@ class Unpack(Node):
         :py:class:`~morphocut.stream.Pack`
     """
 
-    def __init__(self, iterable: RawOrVariable[Iterable]):
+    def __init__(self, collection: RawOrVariable[Collection]):
         super().__init__()
-        self.iterable = iterable
+        self.collection = collection
 
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: Stream):
         """Transform a stream."""
 
         with closing_if_closable(stream):
+            stream_estimator = StreamEstimator()
             for obj in stream:
-                iterable = self.prepare_input(obj, "iterable")
-
-                for value in iterable:
-                    yield self.prepare_output(obj.copy(), value)
+                collection = tuple(self.prepare_input(obj, "collection"))
+                with stream_estimator.incoming_object(
+                    obj.n_remaining_hint, len(collection)
+                ):
+                    for value in collection:
+                        yield self.prepare_output(
+                            obj.copy(), value, n_remaining_hint=stream_estimator.emit()
+                        )
 
 
 @ReturnOutputs
@@ -276,16 +284,26 @@ class Pack(Node):
         # Mess with self.outputs
         self.outputs = [Variable(v.name, self) for v in self.variables]
 
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: Stream):
+        stream_estimator = StreamEstimator()
+
         while True:
             packed = list(itertools.islice(stream, self.size))
 
             if not packed:
                 break
 
-            packed_values = tuple(tuple(o[v] for o in packed) for v in self.variables)
+            with stream_estimator.incoming_object(
+                packed[0].n_remaining_hint, n_consumed=len(packed)
+            ):
 
-            yield self.prepare_output(packed[0], *packed_values)
+                packed_values = tuple(
+                    tuple(o[v] for o in packed) for v in self.variables
+                )
+
+                yield self.prepare_output(
+                    packed[0], *packed_values, n_remaining_hint=stream_estimator.emit()
+                )
 
 
 class _Predicate:
@@ -296,7 +314,6 @@ class _Predicate:
         return obj[self.variable]
 
 
-@ReturnOutputs
 class Filter(Node):
     """
     |stream| Filter objects in the :py:obj:`~morphocut.core.Stream`.
@@ -339,13 +356,17 @@ class Filter(Node):
         else:
             self.predicate = predicate
 
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: Stream):
         with closing_if_closable(stream):
+            stream_estimator = StreamEstimator()
             for obj in stream:
-                if not self.predicate(obj):
-                    continue
+                with stream_estimator.incoming_object(obj.n_remaining_hint):
 
-                yield obj
+                    if not self.predicate(obj):
+                        continue
+
+                    obj.n_remaining_hint = stream_estimator.emit()
+                    yield obj
 
 
 @ReturnOutputs
@@ -364,7 +385,7 @@ class FilterVariables(Node):
             for v in variables
         }
 
-    def transform_stream(self, stream):
+    def transform_stream(self, stream: Stream):
         with closing_if_closable(stream):
             for obj in stream:
                 yield StreamObject({k: v for k, v in obj.items() if k in self.keys})